♻️ refactor(api): consolidate utility imports into a single index file
