+++ apps/api/src/application/use-cases/email-verification/email-verification-request.usecase.ts
@@ -1,6 +1,5 @@
 import { emailVerificationCodeExpiresSpan } from "../../../common/constants";
-import { generateRandomString } from "../../../common/utils/generate-random-value";
-import { ulid } from "../../../common/utils/ulid";
+import { generateRandomString, ulid } from "../../../common/utils";
 import type { User } from "../../../domain/user";
 import type { IEmailVerificationCodeRepository } from "../../../interface-adapter/repositories/email-verification-code";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
