+++ apps/api/src/common/utils/convert-redirectable-mobile-scheme.ts
@@ -1,6 +1,4 @@
-const convertRedirectableMobileScheme = (url: string | URL): string => {
+export const convertRedirectableMobileScheme = (url: string | URL): string => {
 	const _url = typeof url === "string" ? new URL(url).toString() : url.toString();
 	return _url.replace(/\/\/\//g, "//");
 };
-
-export { convertRedirectableMobileScheme };
