--- apps/api/src/common/utils/convert-redirectable-mobile-scheme/index.ts
@@ -1,1 +1,0 @@
-export { convertRedirectableMobileScheme } from "./convert-redirectable-mobile-scheme";
