--- apps/api/src/common/utils/generate-random-value/index.ts
@@ -1,1 +1,0 @@
-export { random, generateRandomIntegerNumber, generateRandomString } from "./generate-random-value";
