--- apps/api/src/common/utils/get-ip/index.ts
@@ -1,1 +1,0 @@
-export { getIP } from "./get-ip";
