+++ apps/api/src/common/utils/read-bearer-token.ts
