--- apps/api/src/common/utils/read-bearer-token/index.ts
@@ -1,1 +1,0 @@
-export { readBearerToken } from "./read-bearer-token";
