+++ apps/api/src/common/utils/read-session-cookie.ts
