--- apps/api/src/common/utils/read-session-cookie/index.ts
@@ -1,1 +1,0 @@
-export { readSessionCookie } from "./read-session-cookie";
