+++ apps/api/src/common/utils/tests/convert-redirectable-mobile-scheme.test.ts
