+++ apps/api/src/common/utils/time-span.ts
