--- apps/api/src/common/utils/time-span/index.ts
@@ -1,1 +1,0 @@
-export { TimeSpan, type TimeSpanUnit } from "./time-span";
