--- apps/api/src/common/utils/ulid/index.ts
@@ -1,1 +1,0 @@
-export { ulid } from "./ulid";
