+++ apps/api/src/modules/auth-guard/auth-guard.plugin.ts
@@ -1,7 +1,6 @@
 import { ValidateSessionUseCase } from "../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../common/constants";
-import { readBearerToken } from "../../common/utils/read-bearer-token";
-import { readSessionCookie } from "../../common/utils/read-session-cookie";
+import { readBearerToken, readSessionCookie } from "../../common/utils";
 import type { Session } from "../../domain/session";
 import type { User } from "../../domain/user";
 import { DrizzleService } from "../../infrastructure/drizzle";
